rename gem_resource to gemd_resource
