+++ src/citrine/resources/gemd_resource.py
